db migrated and added the relations

+++ db/migrate/20130529235151_create_rewards.rb
@@ -3,7 +3,7 @@ class CreateRewards < ActiveRecord::Migration
     create_table :rewards do |t|
       t.text :description
       t.integer :project_id
-      t.numeric :minimum_value
+      t.integer :minimum_value
       t.string :maximum_backers
 
       t.timestamps
